pyproject.toml
[tool.poetry]
name = "repl_python3_Hackster"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube-dl = "^2021.6.6"
pyglet = "^1.5.18"
opencv-python = "^4.5.2"
speedtest-cli = "^2.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
